java.bbclass: java missing classes during compile
oe_makeclasspath is used to define the classpath argument passed to
javac for a number of packages. The current behaviour takes a number
args specifying jar files to look for and if the files exist, it adds
it to a specified variable.

On some builds, oe_makeclasspath cannot find the jar file.

oe_makeclasspath used an invalid test to see the package is native.
(testing if the build arch and package arch are the same, if so, native)
We test to see if PN matches *-native or *-nativesdk instead.

(LOCAL REV: NOT UPSTREAM) -- Sent to oe-devel on 20151224

Signed-off-by: Amy Fong <amy.fong@windriver.com>
Signed-off-by: Jackie Huang <jackie.huang@windriver.com>